requirements.txt
chainlit==0.4.1
langchain==0.0.208
openai==0.27.8
pinecone_client==2.2.2
python-dotenv==1.0.0
